My first commit. to this great kernel!
Added my tweaks to build script.
Added tune to arch/arm/mm/cache-v7.S to reduce not needed error check, to boost performance.
Tuned block/sio-iosched.c as instructed for flash devices.
Added my fast commands .sh files.
Added latest commits from origin kernel.
Added latest drivers/cpufreq/cpufreq_interactive.c from V4 iCS source
Tuned drivers/cpufreq/cpufreq_ondemand.c and added tuning via menu.
Added improved ZRAM and activated in settings + swap.
Tuned RAM settings.
Removed DEBUG from settings, it's waste of resources for daily use.
I will set one more kernel with DEBUG to debug when needed.

+++ build_kernel.sh
@@ -2,6 +2,7 @@
 export KERNELDIR=`readlink -f .`
 export INITRAMFS_SOURCE=`readlink -f $KERNELDIR/../initramfs3`
 export PARENT_DIR=`readlink -f ..`
+export USE_SEC_FIPS_MODE=true
 
 if [ "${1}" != "" ];then
   export KERNELDIR=`readlink -f ${1}`
@@ -11,48 +12,61 @@ INITRAMFS_TMP="/tmp/initramfs-source"
 
 if [ ! -f $KERNELDIR/.config ];
 then
-  make siyah_defconfig
+  make dorimanx_defconfig
 fi
 
 . $KERNELDIR/.config
 
-rm -rf $INITRAMFS_TMP
-rm -rf $INITRAMFS_TMP.cpio
-mkdir $INITRAMFS_TMP
+#remove previous zImage files
+if [ -e $KERNELDIR/zImage ]; then
+rm $KERNELDIR/zImage
+rm $KERNELDIR/arch/arm/boot/zImage
+fi
 
 export ARCH=arm
-export CROSS_COMPILE=$PARENT_DIR/toolchain/bin/arm-eabi-
+export CROSS_COMPILE=$PARENT_DIR/toolchain/bin/arm-none-eabi-
 
 cd $KERNELDIR/
-nice -n 10 make -j4 || exit 1
+nice -n 10 make -j8 || exit 1
 
+#remove previous initramfs files
+if [ -d $INITRAMFS_TMP ]; then
 rm -rf $INITRAMFS_TMP
+fi
+if [ -e $INITRAMFS_TMP.cpio ]; then
+rm -rf $INITRAMFS_TMP.cpio
+fi
+#copy initramfs files to tmp directory
 cp -ax $INITRAMFS_SOURCE $INITRAMFS_TMP
+#clear git repositories in initramfs
+if [ -d $INITRAMFS_TMP/.git ]; then
 find $INITRAMFS_TMP -name .git -exec rm -rf {} \;
+fi
+#remove empty directory placeholders
 find $INITRAMFS_TMP -name EMPTY_DIRECTORY -exec rm -rf {} \;
+#remove mercurial repository
+if [ -d $INITRAMFS_TMP/.hg ]; then
 rm -rf $INITRAMFS_TMP/.hg
+fi
+#copy modules into initramfs
 mkdir -p $INITRAMFS/lib/modules
 find -name '*.ko' -exec cp -av {} $INITRAMFS_TMP/lib/modules/ \;
 
-cd $INITRAMFS_TMP
-find | fakeroot cpio -H newc -o > $INITRAMFS_TMP.cpio 2>/dev/null
-ls -lh $INITRAMFS_TMP.cpio
-cd -
+nice -n 10 make -j3 zImage CONFIG_INITRAMFS_SOURCE="$INITRAMFS_TMP" || exit 1
 
-nice -n 10 make -j3 zImage CONFIG_INITRAMFS_SOURCE="$INITRAMFS_TMP.cpio" || exit 1
-
-cd ../payload
-rm -f ../payload.cpio
-find | fakeroot cpio -H newc -o > ../payload.cpio
-cd -
-
-cd ../recovery
-rm -f ../recovery.cpio
-rm -f ../recovery.cpio.xz
-find | fakeroot cpio -H newc -o > ../recovery.cpio
-cd -
-xz -1 ../recovery.cpio
-
+if [ -e $KERNELDIR/arch/arm/boot/zImage ]; then
 cp $KERNELDIR/arch/arm/boot/zImage zImage
 #$KERNELDIR/mkshbootimg.py $KERNELDIR/zImage $KERNELDIR/arch/arm/boot/zImage $KERNELDIR/../payload.cpio $KERNELDIR/../recovery.cpio.xz
+
+#Copy all needed to ready kernel folder.
+cp $KERNELDIR/.config $KERNELDIR/arch/arm/configs/dorimanx_defconfig
+cp $KERNELDIR/.config $KERNELDIR/READY/
+rm $KERNELDIR/READY/boot/zImage
+rm $KERNELDIR/READY/Kernel_Dorimanx-SGII-ICS-V*
+cp $KERNELDIR/zImage /$KERNELDIR/READY/boot/
+cd $KERNELDIR/READY/
+zip -r Kernel_Dorimanx-SGII-ICS-V1.`date +"%H-%M-%d%m%y"`.zip .
+else
+echo "Kernel STUCK in BUILD! no zImage exist"
+fi
